Move users.json and markdown-badges.json to data folder in src/assets.
